pyproject.toml
[project]
name = "studio-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.3.1",
    "apscheduler>=3.11.0",
    "asyncpg>=0.30.0",
    "backoff>=2.2.1",
    "orjson>=3.11.1",
    "pydantic>=2.11.7",
    "rich>=13.0.0",
    "pydantic[email]",
    "litellm>=1.75.3",
    "fastapi-sso>=0.18.0",
    "anyio>=4.6",
    "mcp>=1.13.0",
    "httpx>=0.27.0",
    "openai-agents[litellm]>=0.3.3",
    "openai>=1.0.0",
    "sensespace-did>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
